sync: the CRDT engine moves to the kernel
PR-4 of the spaces collaboration work, and the point at which the engine stops
being slides' and starts being the platform's.

It could move because it no longer knows what a slide is. A document shape is
two property names injected at construction, so the algebra — composite node
keys, fractional position keys, per-(node,key) registers, the token RGA over
`html` — carries no app vocabulary at all. Moving it BEFORE that would have
parked `doc.slides` and `sl.elements` inside kernel/src for a PR, against
kernel/README.md's own rule that the kernel never sees an app's content shape.

kernel/src/sync/crdt.ts holds `SyncEngine`, `DocShape` and `shape()`. It
declares the three types it needs structurally rather than importing an app
model, because the dependency direction is app → kernel only. What the engine
actually requires of a document is a container of parented nodes each with an
`id`, under two keys the shape names — nothing more.

slides/src/sync/crdt.ts is now a FACADE, and stays one. Every consumer that
imported the engine from that path still can: session.ts, online.ts,
model.ts's type reach, three scripts, and server/guestbook-daemon, which
bundles `SyncState` from there through wrangler. That last one does not fail
loudly on a missing export — the guestbook just stops archiving — so the
facade's completeness is verified at runtime (15 names) and by bundling the
daemon exactly as wrangler does.

Docs follow the code: PLATFORM §9 and PARALLEL-WORK's kernel zone now name
kernel/src/sync/crdt.ts, and say the rest of slides/src/sync/ (session,
transport, blobs) is still slides-shaped and still serialized. kernel/README.md
adds the engine to its inventory and to its frozen-contracts list, with the
sentence that matters: every byte the engine mints is shipped-file
compatibility surface.

DECISIONS.md carries the ruling — shape at construction, never serialized,
never on the wire; no default on the argument; skipDoc derived; and the wire
vocabulary frozen for every app, which must stay settled before spaces collab
reaches a user, because after that spaces files carry it permanently.

Verified: equivalence rig LIVE and byte-identical at CI settings (a facade over
the live engine compared against the frozen baseline is a genuine comparison —
the case checked when the baseline was frozen); shape seam 14/14; test-sync.ts
46408 checks; test-crdt-delprop 22/22; release rig 40/40; slides and kernel
typecheck standalone; both shells build and pass the splice gate; spaces
unaffected at 97.5% of its ceiling; slides shell 677,668 B (+72 B, the facade).
